+ 5.1.2.0 + Minor housekeeping. + Finalized SuperGump animation engine. + Refactor SuperGumpLayout to back more commonly used Action delegates without arguments. + Implemented use of object pools for common collections in SuperGumps. + Add support for SuperGump to allow multiple instances to be open at once. + Fix potential iterator issues when calling static SuperGump enumeration methods. + Fix issue with restocking DynamicBuyInfo for AdvancedVendors. + Update custom shrink system support for AdvancedVendors. + Cast usages of TileFlags to UInt64 in ArtExtUtility for compatibility. + Add extension method for entry insertion at specific Dictionary index.

+++ Extensions/Ultima/ArtExt.cs
@@ -463,9 +463,9 @@ namespace Ultima
 			foreach (var e in mcl.List //
 								 .OrderBy(o => ((o.m_OffsetX * mcl.Height) + o.m_OffsetY) * 2)
 								 .ThenBy(zt => zt.m_OffsetZ)
-								 .ThenByDescending(zt => (zt.m_Flags & (long)Server.TileFlag.Surface) != 0)
-								 .ThenByDescending(zt => (zt.m_Flags & (long)Server.TileFlag.Wall) != 0)
-								 .ThenBy(zt => (zt.m_Flags & (long)Server.TileFlag.Roof) != 0)
+								 .ThenByDescending(zt => ((ulong)zt.m_Flags & (ulong)Server.TileFlag.Surface) != 0)
+								 .ThenByDescending(zt => ((ulong)zt.m_Flags & (ulong)Server.TileFlag.Wall) != 0)
+								 .ThenBy(zt => ((ulong)zt.m_Flags & (long)Server.TileFlag.Roof) != 0)
 								 .ThenBy(zt => Server.TileData.ItemTable[zt.m_ItemID].CalcHeight))
 			{
 				yield return e;
